clinical_service/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0